Moved test types to a separate namespace.
