+++ Src/SemanticComparisonUnitTest/SemanticComparisonUnitTest.csproj
@@ -74,15 +74,15 @@
     <Compile Include="MemberInfoNameComparerTest.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ProxyCreationExceptionTest.cs" />
-    <Compile Include="PublicSealedType.cs" />
+    <Compile Include="TestTypes\PublicSealedType.cs" />
     <Compile Include="SemanticComparerTest.cs" />
-    <Compile Include="TypeOverridingGetHashCode.cs" />
-    <Compile Include="TypeWithDifferentParameterTypesAndProperties.cs" />
-    <Compile Include="TypeWithIdenticalParameterTypesAndProperties.cs" />
-    <Compile Include="TypeWithIncompatibleAndCompatibleConstructor.cs" />
-    <Compile Include="TypeWithPrivateDefaultCtorAndOtherCtor.cs" />
-    <Compile Include="TypeWithPublicFieldsAndProperties.cs" />
-    <Compile Include="TypeWithUnorderedProperties.cs" />
+    <Compile Include="TestTypes\TypeOverridingGetHashCode.cs" />
+    <Compile Include="TestTypes\TypeWithDifferentParameterTypesAndProperties.cs" />
+    <Compile Include="TestTypes\TypeWithIdenticalParameterTypesAndProperties.cs" />
+    <Compile Include="TestTypes\TypeWithIncompatibleAndCompatibleConstructor.cs" />
+    <Compile Include="TestTypes\TypeWithPrivateDefaultCtorAndOtherCtor.cs" />
+    <Compile Include="TestTypes\TypeWithPublicFieldsAndProperties.cs" />
+    <Compile Include="TestTypes\TypeWithUnorderedProperties.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\SemanticComparison\SemanticComparison.csproj">
