+++ Src/SemanticComparisonUnitTest/TestTypes/PublicSealedType.cs
@@ -1,4 +1,4 @@
-﻿namespace Ploeh.SemanticComparison.UnitTest
+﻿namespace Ploeh.SemanticComparison.UnitTest.TestTypes
 {
     public sealed class PublicSealedType
     {
