+++ Src/SemanticComparisonUnitTest/TestTypes/TypeOverridingGetHashCode.cs
@@ -1,4 +1,4 @@
-﻿namespace Ploeh.SemanticComparison.UnitTest
+﻿namespace Ploeh.SemanticComparison.UnitTest.TestTypes
 {
     public class TypeOverridingGetHashCode
     {
